fix: change gpgconf path to AppData
the SO answer said "for older versions". my current pc i write this on has gpg --version "2.4.0" which is i'd think new but only using the appdata path did the value get shown to have changed.
finally, i can rest and only need to enter it one time a day